Adding decorators
Adding some more decorators and putting them in a separate file for reusability for now..
